fix: use schema to get proper data type for fy
