+++ backend/ops_api/ops/resources/portfolio_funding_summary.py
@@ -4,6 +4,7 @@ from models.base import BaseModel
 from ops_api.ops.auth.auth_types import Permission, PermissionType
 from ops_api.ops.auth.decorators import is_authorized
 from ops_api.ops.base_views import BaseItemAPI
+from ops_api.ops.schemas.portfolio_funding_summary import RequestSchema
 from ops_api.ops.utils.fiscal_year import get_current_fiscal_year
 from ops_api.ops.utils.portfolios import get_total_funding
 from ops_api.ops.utils.response import make_response_with_headers
@@ -18,11 +19,9 @@ class PortfolioFundingSummaryItemAPI(BaseItemAPI):
         """
         /portfolio-funding-summary/<int:id>
         """
-        fiscal_year = request.args.get("fiscal_year")
-
-        if not fiscal_year:
-            fiscal_year = get_current_fiscal_year()
+        schema = RequestSchema()
+        data = schema.load(request.args)
 
         portfolio = self._get_item(id)
-        portfolio_funding_summary = get_total_funding(portfolio, fiscal_year)
+        portfolio_funding_summary = get_total_funding(portfolio, data.get("fiscal_year", get_current_fiscal_year()))
         return make_response_with_headers(portfolio_funding_summary)
