+++ backend/ops_api/ops/schemas/portfolio_funding_summary.py
@@ -1,0 +1,5 @@
+from marshmallow import Schema, fields
+
+
+class RequestSchema(Schema):
+    fiscal_year = fields.Integer(allow_none=True)
